Moved multiply_two_numbers.py from unit_tests to the parent folder
Closes #8.
